reflink: align data on block size on archive create
Add a --reflink option which, on archive creation, aligns file data
on a 4k block boundary. This is done by padding the header with a
standard "comment" pax record, which will be ignored on extraction.

This will be used when extracting files using reflinks, as one of the
requirements is that the data is aligned on a 4k boundary.