reflink: add support for reflink when extracting
Add support for copy-on-write archive extraction via the --reflink option.
This uses the filesystem FICLONERANGE ioctl to a lightweight copy,
similarly to what `cp --reflink=auto` do.
This has two advantages:
first, the extraction is faster because the only IO performed is to read
the header entries from the archive and create destination inodes:

	$ time tar xf linux-6.11.5.tar

	real	0m14.329s
	user	0m0.225s
	sys	0m3.359s

	$ time tar xf linux-6.11.5.tar --reflink

	real	0m1.703s
	user	0m0.180s
	sys	0m1.298s

second, because of the block sharing, the extracted files don't take
more space than the original archive alone:

	$ df -h .
	Filesystem      Size  Used Avail Use% Mounted on
	/dev/vda3        19G   15G  3.5G  81% /home
	$ tar xf linux-6.11.5.tar
	$ df -h .
	Filesystem      Size  Used Avail Use% Mounted on
	/dev/vda3        19G   15G  3.2G  83% /home
	$ rm -rf linux-6.11.5
	$ tar xf linux-6.11.5.tar --reflink
	$ df -h .
	Filesystem      Size  Used Avail Use% Mounted on
	/dev/vda3        19G   15G  3.5G  81% /home

If some some reason the reflink fails (unsupported by the filesystem,
different mountpoint, etc.) a regular copy is made as fallback.